+++ wasm-calc10/src/Calc/Wasm/FromExpr/Expr.hs
@@ -92,24 +92,6 @@ fromMatch expr _pats = do
   -- TODO: actually make this do something good
   pure wasmExpr
 
--- | we use a combination of the value and the type
-fromPrim :: (MonadError FromWasmError m) => Type ann -> Prim -> m WasmPrim
-fromPrim _ (PBool b) = pure $ WPBool b
-fromPrim (TPrim _ TFloat32) (PFloatLit f) =
-  pure $ WPFloat32 (realToFrac f)
-fromPrim (TPrim _ TFloat64) (PFloatLit f) =
-  pure $ WPFloat64 f
-fromPrim (TPrim _ TInt8) (PIntLit i) =
-  pure (WPInt32 (fromIntegral i))
-fromPrim (TPrim _ TInt16) (PIntLit i) =
-  pure (WPInt32 (fromIntegral i))
-fromPrim (TPrim _ TInt32) (PIntLit i) =
-  pure (WPInt32 (fromIntegral i))
-fromPrim (TPrim _ TInt64) (PIntLit i) =
-  pure (WPInt64 (fromIntegral i))
-fromPrim ty prim =
-  throwError $ PrimWithNonNumberType prim (void ty)
-
 fromExprWithDrops ::
   ( MonadError FromWasmError m,
     MonadState FromExprState m,
