+++ wasm-calc10/test/Test/Wasm/FromWasmSpec.hs
@@ -11,6 +11,7 @@ import Calc.Wasm.FromExpr.Drops
     typeToDropPaths,
   )
 import Calc.Wasm.FromExpr.Helpers (monomorphiseTypes)
+import Calc.Wasm.FromExpr.Patterns.Predicates
 import Calc.Wasm.ToWasm.Types
 import Control.Monad (void)
 import Data.Foldable (traverse_)
@@ -128,3 +129,30 @@ spec = do
               typeToDropPaths (unsafeTy tyString) id `shouldBe` paths
         )
         testVals
+
+    fdescribe "Predicates" $ do
+      describe "predicateToWasm" $ do
+        let testVals =
+              [ ( Equals [] tyBool (PBool True),
+                  WVar 100,
+                  WInfix I8 OpEquals (WPrim (WPBool True)) (WVar 100)
+                ),
+                ( Equals [] tyInt32 (PIntLit 42),
+                  WVar 100,
+                  WInfix I8 OpEquals (WPrim (WPInt32 42)) (WVar 100)
+                ),
+                ( Equals [(tyInt32, 0)] tyInt32 (PIntLit 42),
+                  WVar 100,
+                  WInfix I8 OpEquals (WPrim (WPInt32 42)) (WTupleAccess I32 (WVar 100) 0)
+                ),
+                ( Equals [(tyInt32, 0), (tyInt64, 1)] tyInt32 (PIntLit 42),
+                  WVar 100,
+                  WInfix I8 OpEquals (WPrim (WPInt32 42)) (WTupleAccess I32 (WTupleAccess I64 (WVar 100) 1) 0)
+                )
+              ]
+        traverse_
+          ( \(predicate, val, expected) ->
+              it (show predicate) $ do
+                predicateToWasm @_ @() predicate val `shouldBe` Right expected
+          )
+          testVals
