+++ wasm-calc10/test/Test/Wasm/WasmSpec.hs
@@ -312,7 +312,7 @@ spec = do
               )
             ]
 
-      fdescribe "From expressions" $ do
+      describe "From expressions" $ do
         traverse_ testWithInterpreter testVals
 
     describe "Run tests" $ do
